Disable text selection of UI elements
Trying to move a slider while body elements are selected can cause the slider to lock on chromium-based browsers. Since none of the body needs to be copied, selection can be disabled entirely.

My original idea was just to de-select on sliderChange, I've included that option if more fine-grained control of user selection is needed.